chore: split `collections.rs` into two
